+++ java/server/src/org/openqa/selenium/grid/commands/InfoCommand.java
@@ -24,6 +24,7 @@ import com.google.common.io.Resources;
 
 import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.PrintStream;
 import java.io.StringReader;
 import java.nio.charset.StandardCharsets;
 import org.openqa.selenium.cli.CliCommand;
@@ -32,7 +33,6 @@ import org.openqa.selenium.grid.server.HelpFlags;
 
 import java.io.PrintWriter;
 import java.util.Collections;
-
 
 @AutoService(CliCommand.class)
 public class InfoCommand implements CliCommand {
@@ -45,7 +45,7 @@ public class InfoCommand implements CliCommand {
     return "Prints information for commands and topics.";
   }
 
-  public Executable configure(String... args) {
+  public Executable configure(PrintStream out, PrintStream err, String... args) {
     HelpFlags help = new HelpFlags();
     InfoFlags topic = new InfoFlags();
 
@@ -59,12 +59,12 @@ public class InfoCommand implements CliCommand {
       try {
         commander.parse(args);
       } catch (ParameterException e) {
-        System.err.println(e.getMessage());
+        err.println(e.getMessage());
         commander.usage();
         return;
       }
 
-      if (help.displayHelp(commander, System.out)) {
+      if (help.displayHelp(commander, out)) {
         return;
       }
 
@@ -92,11 +92,11 @@ public class InfoCommand implements CliCommand {
       String path = getClass().getPackage().getName().replaceAll("\\.", "/") + "/" + toDisplay;
       String content = readContent(path);
 
-      PrintWriter out = new WrappedPrintWriter(System.out, 72, 0);
+      PrintWriter outWriter = new WrappedPrintWriter(out, 72, 0);
 
-      out.printf("\n%s\n%s\n\n", title, String.join("", Collections.nCopies(title.length(), "=")));
-      out.print(content);
-      out.println("\n\n");
+      outWriter.printf("\n%s\n%s\n\n", title, String.join("", Collections.nCopies(title.length(), "=")));
+      outWriter.print(content);
+      outWriter.println("\n\n");
     };
   }