+++ javascript/node/selenium-webdriver/ie.js
@@ -219,18 +219,21 @@ class Options extends Capabilities {
     return this;
   }
 
-  /**
-   * Specifies command-line switches to use when launching Internet Explorer.
-   * This is only valid when used with {@link #forceCreateProcessApi}.
-   *
-   * @param {...(string|!Array.<string>)} args The arguments to add.
-   * @return {!Options} A self reference.
-   */
-  addArguments(...args) {
-    let current = this.get(Key.BROWSER_COMMAND_LINE_SWITCHES) || [];
-    this.options_[Key.BROWSER_COMMAND_LINE_SWITCHES] = current.concat.apply(current, args);
-    return this;
-  }
+ /**
+  * Specifies command-line switches to use when launching Internet Explorer.
+  * This is only valid when used with {@link #forceCreateProcessApi}.
+  *
+  * @param {...(string|!Array.<string>)} args The arguments to add.
+  * @return {!Options} A self reference.
+  */
+
+ addArguments(...args) {
+   let current = this.options_[Key.BROWSER_COMMAND_LINE_SWITCHES] || [];
+   if(typeof current == 'string')
+     current = current.split(" ");
+   this.options_[Key.BROWSER_COMMAND_LINE_SWITCHES] = current.concat(args).join(" ");
+   return this;
+ }
 
   /**
    * Configures whether proxies should be configured on a per-process basis. If
@@ -388,22 +391,22 @@ function createServiceFromCapabilities(capabilities) {
         'ensure it can be found on your system PATH.');
   }
 
-  var args = [];
-  if (capabilities.has(Key.HOST)) {
-    args.push('--host=' + capabilities.get(Key.HOST));
-  }
-  if (capabilities.has(Key.LOG_FILE)) {
-    args.push('--log-file=' + capabilities.get(Key.LOG_FILE));
-  }
-  if (capabilities.has(Key.LOG_LEVEL)) {
-    args.push('--log-level=' + capabilities.get(Key.LOG_LEVEL));
-  }
-  if (capabilities.has(Key.EXTRACT_PATH)) {
-    args.push('--extract-path=' + capabilities.get(Key.EXTRACT_PATH));
-  }
-  if (capabilities.get(Key.SILENT)) {
-    args.push('--silent');
-  }
+ var args = [];
+ if (capabilities.has(Key.HOST)) {
+   args.push('--host=' + capabilities.get(Key.HOST));
+ }
+ if (capabilities.has(Key.LOG_FILE)) {
+   args.push('--log-file=' + capabilities.get(Key.LOG_FILE));
+ }
+ if (capabilities.has(Key.LOG_LEVEL)) {
+   args.push('--log-level=' + capabilities.get(Key.LOG_LEVEL));
+ }
+ if (capabilities.has(Key.EXTRACT_PATH)) {
+   args.push('--extract-path=' + capabilities.get(Key.EXTRACT_PATH));
+ }
+ if (capabilities.get(Key.SILENT)) {
+   args.push('--silent');
+ }
 
   var port = portprober.findFreePort();
   return new remote.DriverService(exe, {